backend/requirements.txt
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djoser==2.3.1
fdb==2.0.2
flake8==7.2.0
flake8-broken-line==1.0.0
flake8-isort==6.1.2
flake8-return==1.2.0
gunicorn==23.0.0
isort==6.0.1
pep8-naming==0.14.1
psycopg==3.2.6
python-dotenv==1.1.0
